Fix TypeError when mixing float and Decimal in calc_transactions
Convert all amounts to float before processing to prevent TypeError when
groupby().sum() tries to add float and Decimal types together at line 312.

This ensures type consistency throughout the calculation:
- Convert entire amount column to float using astype(float) at line 306
- Remove redundant float() conversion for expenses since all are now float

Fixes: TypeError: unsupported operand type(s) for +: 'float' and 'decimal.Decimal'

+++ app/cashflow.py
@@ -302,9 +302,11 @@ def calc_transactions(balance, total):
     # for schedules marked as expenses, make the value negative for the sum
     # Create a copy to avoid modifying during iteration
     df = df.copy()
+    # Convert all amounts to float to avoid Decimal/float mixing
+    df['amount'] = df['amount'].astype(float)
     for idx in df.index:
         if df.loc[idx, 'type'] == 'Expense':
-            df.loc[idx, 'amount'] = float(df.loc[idx, 'amount']) * -1
+            df.loc[idx, 'amount'] = df.loc[idx, 'amount'] * -1
 
     # group total transactions by date and sum the amounts for each date
     df = df.groupby("date")['amount'].sum().reset_index()
